safari-600.7-branch/LayoutTests/js/dom/removing-Cf-characters-expected.txt

This test checks that BOM is treated as whitespace as required by ES5, but other Cf characters are not, despite what ECMA-262v3 says, see <https://bugs.webkit.org/show_bug.cgi?id=4931>.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS escape(testString) is '%uFEFF%u200F%u200E%AD%u2062%u200D%u200C%u200B'
PASS escape(testString2) is '%uFEFF%u200F%u200E%AD%u2062%u200D%u200C%u200B'
PASS 1 is 1
PASS eval('"﻿"').length is 1
PASS eval('"﻿"').charCodeAt(0) is 0xFEFF
PASS +﻿+1 /* BOM between the +'s */ is 1
PASS successfullyParsed is true

TEST COMPLETE
